Added style to project
