+++ package.json
@@ -15,8 +15,12 @@
     "@babel/core": "^7.22.9",
     "@babel/plugin-transform-modules-commonjs": "^7.22.5",
     "@babel/preset-env": "^7.22.9",
+    "babel-eslint": "^10.1.0",
     "babel-loader": "^9.1.3",
     "css-loader": "^6.8.1",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.28.0",
     "esm": "^3.2.25",
     "express": "^4.18.2",
     "file-loader": "^6.2.0",
@@ -24,6 +28,10 @@
     "jest": "^29.6.1",
     "jest-environment-jsdom": "^29.6.1",
     "style-loader": "^3.3.3",
+    "stylelint": "^13.13.1",
+    "stylelint-config-standard": "^21.0.0",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "^3.21.0",
     "url-loader": "^4.1.1",
     "webpack": "^5.88.1",
     "webpack-cli": "^5.1.4",
