+++ src/index.html
@@ -22,7 +22,7 @@
                 </ul>
             </nav>
         </header>
-        <section>
+        <section class="foodSection">
             <div id="mealListContainer">
             <!-- The list will be displayed here -->
           </div>
